Move deployment to file systems
